Instantiating a cube using Raycasting (playerController script)
added serialized variables float shootRate, int shotDist, GameObject cube
added bool isShooting
created a IEnumerator shoot() function
setting isShooting to true;
yield return new waitForSeconds(shootRate); making it wait till when we can shoot again
than if statment using physic.raycast() function that needs a direction and position
inside that function we used the main camera with another function ViewportPointToRay() asking for two point of a Vector 3 .5f , .5f this will instanitate the cube in the middle of the screen. and out hit the that returns the postion of where the ray cast hit and the shoot dist of how far we want the ray cast to go.
called the function in the Update function using the StartCoroutine() function with and if statement to be able to use the left mouse click(Input.GetButton()function) to execute the functions