Centered Cube
when using the ray cast to place cubes on the floor they would be set un the middle of the floor.
To fix this I
first made a Empty Game Object
Next reset its position.
Create a child cube object and position its why to where it is sitting correctly on the floor (y at .5)
than made it a prefab and put that prefab in the players script